requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
line-bot-sdk==3.11.0
google-generativeai==0.8.0
pydantic-settings==2.3.0
redis==5.0.8
aiofiles==24.1.0
